pyproject.toml
[tool.poetry]
name = "wronggpt"
version = "0.1.0"
description = ""
authors = ["Harris Lapiroff <harris@freedom.press>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.2.2"
openai = "^0.27.2"

[tool.poetry.scripts]
bot = "wronggpt.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"